cpp 09

+++ CPP_Module_09/ex02/PmergeMe.hpp
@@ -6,7 +6,7 @@
 /*   By: libacchu <libacchu@students.42wolfsburg    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/14 16:04:13 by libacchu          #+#    #+#             */
-/*   Updated: 2023/03/16 11:32:37 by libacchu         ###   ########.fr       */
+/*   Updated: 2023/03/16 15:01:31 by libacchu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,7 +43,7 @@ class PmergeMe
 		void dequeInsertSort(std::deque<int> &deq, int low, int high);
 		void dequeMergeInsertSort(std::deque<int> &deq, int low, int high);
 		
-		double	get_time_in_ms(void);
+		long double	get_time_in_ms(void);
 
 		void printVector(std::vector<int>& vec);
 };
